Add basic skeleton
This raises the GC threshold to the maximum available so it doesn't
kick often on first load, and then sets it to 1GB after Emacs loads.

It also prevents some filename parsing when loading init files.

Last but not least, it will try and load the file where the custom
variables are stored.